Revert "add a cleanup script"
