+++ action.yml
@@ -8,8 +8,7 @@ inputs:
     default: 0.8.4 # latest
 runs:
   using: "node20"
-  main: "dist/setup/index.js"
-  post: "dist/post/index.js"
+  main: "dist/index.js"
 branding:
   icon: "navigation"
   color: "green"
