+++ package.json
@@ -2,9 +2,9 @@
   "name": "setup-nf-test",
   "version": "0.1.0",
   "description": "",
-  "main": "dist/setup/index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "build": "ncc build src/setup.js -o dist/setup --license licenses.txt && ncc build src/post.js -o dist/post --license licenses.txt",
+    "build": "ncc build index.js --license licenses.txt",
     "prepare": "husky"
   },
   "keywords": [
