--- src/post.js
@@ -1,20 +1,0 @@
-const fs = require("fs").promises
-const os = require("os")
-const path = require("path")
-const { debug } = require("@actions/core")
-
-async function cleanup() {
-  try {
-    const dirPath = path.join(os.homedir(), ".nf-test")
-    await fs.rm(dirPath, { recursive: true })
-    debug(`Directory ${dirPath} has been removed`)
-  } catch (error) {
-    debug(`Failed to remove directory: ${error}`)
-  }
-}
-
-module.exports = cleanup
-
-if (require.main === module) {
-  cleanup()
-}
